+++ guardrails_api/start.sh
@@ -1,1 +1,1 @@
-gunicorn --bind 0.0.0.0:8000 --timeout=5 --threads=10 "guardrails_api.app:create_app()"
+gunicorn --bind 0.0.0.0:8000 --timeout=5 --workers=3 --worker-class=uvicorn.workers.UvicornWorker "guardrails_api.app:create_app()"
