+++ pyproject.toml
@@ -11,13 +11,10 @@ keywords = ["Guardrails", "Guardrails AI", "Guardrails API", "Guardrails API"]
 requires-python = ">= 3.8.1"
 dependencies = [
     "guardrails-ai>=0.5.6",
-    "flask>=3.0.3,<4",
     "Flask-SQLAlchemy>=3.1.1,<4",
-    "Flask-Caching>=2.3.0,<3",
     "Werkzeug>=3.0.3,<4",
     "jsonschema>=4.22.0,<5",
     "referencing>=0.35.1,<1",
-    "Flask-Cors>=4.0.1,<6",
     "boto3>=1.34.115,<2",
     "psycopg2-binary>=2.9.9,<3",
     "litellm>=1.39.3,<2",
@@ -26,8 +23,10 @@ dependencies = [
     "opentelemetry-sdk>=1.0.0,<2",
     "opentelemetry-exporter-otlp-proto-grpc>=1.0.0,<2",
     "opentelemetry-exporter-otlp-proto-http>=1.0.0,<2",
-    "opentelemetry-instrumentation-flask>=0.12b0,<1",
-    "requests>=2.32.3"
+    "opentelemetry-instrumentation-fastapi>=0.47b0",
+    "requests>=2.32.3",
+    "aiocache>=0.11.1",
+    "fastapi",
 ]
 
 [tool.setuptools.dynamic]
