+++ requirements-lock.txt
@@ -28,10 +28,16 @@ frozenlist==1.4.1
 fsspec==2024.6.1
 googleapis-common-protos==1.63.2
 griffe==0.36.9
+<<<<<<< Updated upstream
 grpcio==1.65.1
 guardrails-ai==0.5.7
 guardrails-api-client==0.3.12
 guardrails_hub_types==0.0.4
+=======
+grpcio==1.64.1
+guardrails-ai==0.5.0a2
+guardrails-api-client==0.3.8
+>>>>>>> Stashed changes
 gunicorn==22.0.0
 h11==0.14.0
 httpcore==1.0.5
